# seven_shoppingStore/swagger.yaml
...
openapi: "3.0.0"
info:
  title: Tshirt store API
  description: LCO - a course to create API for ecomm store
  version: 1.1.0
  contact:
    email: hitesh@lco.dev
    url: "https://lco.dev"

# servers:
#   - url: "https://localhost:4000/api/v1"
#     description: for local host - secure
#   - url: "http://localhost:4000/api/v1"
#     description: for local host - regular
servers:
  - url: "{protocol}://{url}/api/{version}"
    description: for local host 
    variables:
      version:
        enum:
          - v1
          - v2
        default: v1
      url:
        enum:
          - localhost:4000
          - lcotshirtstore.herokuapp.com
        default: lcotshirtstore.herokuapp.com
      protocol:
        enum:
          - http
          - https
        default: https
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
    BearerAuth:
      type: http
      scheme: bearer
paths:
...
  /order/{id}:
    get:
      tags:
        - Order
      summary: get details about a order with order id
      parameters:
        - name: id
          in: path
          required: true
          description: id of the order
          schema:
            type: string
      
      requestBody:
        
      responses:
        200:
          description: All good success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  order:
                    type: object
                    properties:
                      _id:
                        type: string
                      user:
                        type: object
                        properties:
                          name:
                            type: string
                          email:
                            type: string
                          id:
                            type: string
                      orderStatus:
                        type: string
                      createdAt:
                        type: string
                      shippingInfo:
                        type: object
                        properties:
                          address:
                            type: string
                          city: 
                            type: string
                          phoneNo: 
                            type: string
                          postalCode: 
                            type: string
                          state: 
                            type: string
                          country: 
                            type: string
                      paymentInfo:
                        type: object
                        properties:
                          id:
                            type: string
                      taxAmount:
                        type: number
                      shippingAmount:
                        type: number
                      totalAmount:
                        type: number
                      orderItems:
                        type: array
                        items:
                          type: object
                          properties:
                            name:
                              type: string
                            quantity:
                              type: string
                            image:
                              type: string
                            price:
                              type: number
                            product:
                              type: string
                              description: id of product
                        
        400:
          description: Bad request
        500:
          description: internal server error
  /myorder:
    get:
      tags:
        - Order
      summary: get order of logged in user
      requestBody:
        
      responses:
        200:
          description: All good success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  order:
                    type: object
                    properties:
                      _id:
                        type: string
                      user:
                        type: object
                        properties:
                          name:
                            type: string
                          email:
                            type: string
                          id:
                            type: string
                      orderStatus:
                        type: string
                      createdAt:
                        type: string
                      shippingInfo:
                        type: object
                        properties:
                          address:
                            type: string
                          city: 
                            type: string
                          phoneNo: 
                            type: string
                          postalCode: 
                            type: string
                          state: 
                            type: string
                          country: 
                            type: string
                      paymentInfo:
                        type: object
                        properties:
                          id:
                            type: string
                      taxAmount:
                        type: number
                      shippingAmount:
                        type: number
                      totalAmount:
                        type: number
                      orderItems:
                        type: array
                        items:
                          type: object
                          properties:
                            name:
                              type: string
                            quantity:
                              type: string
                            image:
                              type: string
                            price:
                              type: number
                            product:
                              type: string
                              description: id of product
                        
        400:
          description: Bad request
        500:
          description: internal server error
...